feat: split StablePair logic into two contracts
To enable deployment of the StablePair, it was necessary to split the
logic into two contracts. The second contract contains mint/burn logic,
while the former contains the rest.

This current implementation attempts to perform the minimal changes
necessary. A more advanced implementation could use a fallback function
to reduce the cost of function dispatch on the main contract.